fix stripe accounts migration big int

+++ database/migrations/2020_07_17_074500_create_stripe_accounts_and_events.php
@@ -46,7 +46,7 @@ class CreateStripeAccountsAndEvents extends Migration
             $table->string('email')->nullable();
             $table->json('individual')->nullable();
             $table->json('metadata')->nullable();
-            $table->unsignedInteger('owner_id')->nullable();
+            $table->unsignedBigInteger('owner_id')->nullable();
             $table->boolean('payouts_enabled')->nullable();
             $table->string('refresh_token')->nullable();
             $table->json('requirements')->nullable();
